backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.39.0
sse-starlette==1.8.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
snowflake-connector-python==3.7.0
cryptography==41.0.7